pythonserver/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0 